requirements.txt
fastapi[standard]==0.112.0      # FastAPI framework for building APIs
uvicorn==0.25.0       # ASGI server for running the FastAPI app
pandas==2.1.4         # Data manipulation and analysis
scikit-learn==1.3.2   # Machine learning library for evaluation and prediction
SQLAlchemy==2.0.23    # ORM for database interactions
pymysql==1.1.0        # MySQL database driver (use if you're using MySQL)
python-jose==3.3.0    # For JWT-based authentication
passlib==1.7.4        # Password hashing library
joblib==1.3.2         # Model serialization
joblib==1.3.2         # Model serialization
bcrypt==4.2.1         # Install bycrypt 